requirements/ci.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/ci.txt requirements/base.txt requirements/test-tools.in
#
beautifulsoup4==4.9.1     # via webtest
certifi==2020.6.20        # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
cffi==1.14.0              # via -r requirements/base.txt, cryptography
chardet==3.0.4            # via -r requirements/base.txt, requests
coverage==4.5.4           # via -r requirements/test-tools.in
cryptography==3.0         # via -r requirements/base.txt, django-auth-adfs
cssselect==1.1.0          # via pyquery
django-admin-index==1.4.0  # via -r requirements/base.txt
django-appconf==1.0.4     # via -r requirements/base.txt, django-axes
django-auth-adfs-db==0.2.0  # via -r requirements/base.txt
django-auth-adfs==1.3.1   # via -r requirements/base.txt, django-auth-adfs-db
django-axes==5.4.1        # via -r requirements/base.txt
django-choices==1.7.1     # via -r requirements/base.txt, zgw-consumers
django-compat==1.0.15     # via -r requirements/base.txt, django-hijack, django-hijack-admin
django-hijack-admin==2.1.10  # via -r requirements/base.txt
django-hijack==2.1.10     # via -r requirements/base.txt, django-hijack-admin
django-ipware==3.0.0      # via -r requirements/base.txt, django-axes
django-jenkins==0.110.0   # via -r requirements/test-tools.in
django-ordered-model==3.4.1  # via -r requirements/base.txt, django-admin-index
django-redis==4.12.1      # via -r requirements/base.txt
django-relativedelta==1.1.1  # via -r requirements/base.txt, zgw-consumers
django-rosetta==0.9.4     # via -r requirements/base.txt
django-sendfile==0.3.11   # via -r requirements/base.txt
django-sniplates==0.7.0   # via -r requirements/base.txt
django-solo==1.1.3        # via -r requirements/base.txt, django-auth-adfs-db, zgw-consumers
django-webtest==1.9.7     # via -r requirements/test-tools.in
django==2.2.14            # via -r requirements/base.txt, django-appconf, django-auth-adfs, django-auth-adfs-db, django-axes, django-choices, django-jenkins, django-redis, django-relativedelta, django-rosetta, django-sendfile, django-sniplates, djangorestframework, zgw-consumers
djangorestframework==3.12.2  # via -r requirements/base.txt
elastic-apm==5.8.1        # via -r requirements/base.txt
factory-boy==2.12.0       # via -r requirements/test-tools.in
faker==4.1.1              # via factory-boy
freezegun==0.3.15         # via -r requirements/test-tools.in
gemma-zds-client==0.13.4  # via -r requirements/base.txt, zgw-consumers
idna==2.10                # via -r requirements/base.txt, requests
lxml==4.5.2               # via pyquery
pillow==7.2.0             # via -r requirements/base.txt
polib==1.1.0              # via -r requirements/base.txt, django-rosetta
psycopg2==2.8.5           # via -r requirements/base.txt
pycparser==2.20           # via -r requirements/base.txt, cffi
pyjwt==1.7.1              # via -r requirements/base.txt, django-auth-adfs, gemma-zds-client
pyquery==1.4.1            # via -r requirements/test-tools.in
python-dateutil==2.8.1    # via -r requirements/base.txt, django-relativedelta, faker, freezegun
python-decouple==3.3      # via -r requirements/base.txt
python-dotenv==0.14.0     # via -r requirements/base.txt
pytz==2020.1              # via -r requirements/base.txt, django
pyyaml==5.3.1             # via -r requirements/base.txt, gemma-zds-client
redis==3.5.3              # via -r requirements/base.txt, django-redis
requests-mock==1.8.0      # via -r requirements/test-tools.in
requests==2.24.0          # via -r requirements/base.txt, django-auth-adfs, django-rosetta, gemma-zds-client, requests-mock, zgw-consumers
sentry-sdk==0.16.1        # via -r requirements/base.txt
six==1.15.0               # via -r requirements/base.txt, cryptography, django-choices, django-compat, django-rosetta, freezegun, python-dateutil, requests-mock, webtest
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.3.1           # via -r requirements/base.txt, django
tblib==1.6.0              # via -r requirements/test-tools.in
text-unidecode==1.3       # via faker
uritemplate==3.0.1        # via -r requirements/base.txt
urllib3==1.25.9           # via -r requirements/base.txt, elastic-apm, requests, sentry-sdk
uwsgi==2.0.19.1           # via -r requirements/base.txt
waitress==1.4.4           # via webtest
webob==1.8.6              # via webtest
webtest==2.0.35           # via django-webtest
zgw-consumers==0.13.1     # via -r requirements/base.txt
